Major refactor of resource controller
Abstract the definition of a resource so we can separate static vs
javascript logic.  Add a "resource helper" which is basically a
resource factory responsible to determining what kind of resource
object to create.  Move the shared javascript code out into it's own
bean that can be injected.  Initial (untested) serialization code in
the resources. Renamed and added new exceptions that are using
during resource processing.  Small fix in default html template so
all imports/inlcudes have a leading slash.